pyproject.toml
[project]
name = "capability-isolation"
version = "0.1.0"
description = "Post-hoc capability isolation for trained models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "polars>=1.38.1",
    "rich>=14.3.3",
    "scipy>=1.13.0",
    "umap-learn>=0.5.6",
]
